Bug 813269 - Use double to represent time in Web Audio; r=bzbarsky
UltraBlame original commit: b44293987106239f6e14532780c9da505c3bdce6

+++ content/media/webaudio/AudioParam.h
@@ -102,7 +102,7 @@ public:
 
   
   
-  void SetValueCurveAtTime(JSContext* cx, const Float32Array& aValues, float aStartTime, float aDuration, ErrorResult& aRv)
+  void SetValueCurveAtTime(JSContext* cx, const Float32Array& aValues, double aStartTime, double aDuration, ErrorResult& aRv)
   {
     AudioParamTimeline::SetValueCurveAtTime(detail::FloatArrayWrapper(aValues),
                                             aStartTime, aDuration, aRv);
